+++ src/components/NetworkStylingEdgeByPropertyForm.js
@@ -231,6 +231,10 @@ const NetworkStylingEdgeByPropertyForm = ({
             ) && (
               <Button
                 label={t('delete')}
+                tooltip={t('delete')}
+                tooltipOptions={{
+                  position: 'top'
+                }}
                 className="p-button-warning delete-property-style"
                 icon="pi pi-trash"
                 onClick={() => updateStylingByProperties({
@@ -246,6 +250,10 @@ const NetworkStylingEdgeByPropertyForm = ({
 
           <Button
             label={t('save')}
+            tooltip={t('save')}
+            tooltipOptions={{
+              position: 'top'
+            }}
             icon="pi pi-check"
             className="save-property-style"
             disabled={stylingPropertyObject.styleType === 'stylingEdgeLineStyle' ? false : (!stylingPropertyObject.styleValue