+++ src/utils/networkGraphOptions/checkEdgeVisibility.js
@@ -2,6 +2,8 @@ import {
   LABEL_PROPERTY,
 } from '../../constants/graph'
 import store from '../../store'
+import checkValidProperties from './checkValidProperties'
+import checkVisibilityByProperty from './checkVisibilityByProperty'
 
 /**
  * Check edge visibility
@@ -21,13 +23,29 @@ const checkEdgeVisibility = ({
   } = store.getState()
 
   const {
-    isSubClassEdgeVisible
+    isSubClassEdgeVisible,
+    hiddenEdgesProperties,
   } = graphData[currentGraph]
 
   let isVisible = true
 
   if (!isSubClassEdgeVisible) {
     isVisible = objectPropertiesFromApiBackup[edgeId][LABEL_PROPERTY] !== 'Subclass of'
+  }
+
+  if (isVisible) {
+    if (isVisible) {
+      const properties = checkValidProperties({
+        properties: hiddenEdgesProperties
+      })
+
+      if (properties.length > 0) {
+        isVisible = checkVisibilityByProperty({
+          element: objectPropertiesFromApiBackup[edgeId],
+          properties
+        })
+      }
+    }
   }
 
   if (!isVisible) {